+++ apps/io-func-sign-issuer/src/infra/http/handlers/__test__/get-requests-by-dossier.spec.ts
@@ -34,7 +34,7 @@ describe("GetRequestsByDossierHandler", () => {
     environment: "TEST",
     vatNumber: "14711371128" as NonEmptyString,
     department: "dep1" as NonEmptyString,
-    state: "ACTIVE",
+    status: "ACTIVE",
   };
 
   const makeDossier = (): Dossier => ({