+++ apps/io-func-sign-issuer/src/infra/http/handlers/__test__/validate-document.spec.ts
@@ -27,7 +27,7 @@ const issuer: Issuer = {
   environment: "TEST",
   vatNumber: "14711371128" as NonEmptyString,
   department: "dep1" as NonEmptyString,
-  state: "ACTIVE",
+  status: "ACTIVE",
 };
 
 const mocks = { issuer };