+++ apps/io-func-sign-issuer/src/issuer.ts
@@ -65,7 +65,7 @@ export const ApiKey = t.type({
     type: NonEmptyString,
     institutionId: NonEmptyString,
     supportEmail: EmailString,
-    state: t.union([t.literal("active"), t.literal("inactive")]),
+    status: t.union([t.literal("active"), t.literal("inactive")]),
   }),
 });