+++ apps/io-func-sign-issuer/CHANGELOG.md
@@ -1,5 +1,16 @@
 # io-func-sign-issuer
 
+## 0.17.0
+
+### Minor Changes
+
+- a0cea83: Changed issuer's property from state to status
+
+### Patch Changes
+
+- Updated dependencies [a0cea83]
+  - @io-sign/io-sign@1.1.0
+
 ## 0.16.5
 
 ### Patch Changes